pyproject.toml
[project]
name = "notion-to-hatena-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "notion-client",
    "requests",
    "python-dotenv",
    "PySide6",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pyinstaller",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # Pyflakes
    "W",  # pycodestyle warnings
    "I",  # isort
]

[dependency-groups]
dev = [
    "pyinstaller>=6.17.0",
]

